+++ src/services/index.ts
@@ -3,9 +3,6 @@
 *   Copyright © 2024 NatML Inc. All Rights Reserved.
 */
 
-export * from "./environment"
 export * from "./prediction"
 export * from "./predictor"
-export * from "./storage"
 export * from "./user"
-export * from "./value"