+++ src/services/user.ts
@@ -13,33 +13,6 @@ export interface RetrieveUserInput {
     username: string;
 }
 
-export interface UpdateUserInput {
-    /**
-     * User full name.
-     */
-    name?: string;
-    /**
-     * User bio.
-     */
-    bio?: string;
-    /**
-     * User avatar URL.
-     */
-    avatar?: string;
-    /**
-     * User website.
-     */
-    website?: string;
-    /**
-     * User GitHub handle.
-     */
-    github?: string;
-    /**
-     * Monthly spend limit.
-     */
-    spendLimit?: number;
-}
-
 export class UserService {
 
     private readonly client: GraphClient;
@@ -60,24 +33,6 @@ export class UserService {
                 user (input: $input) {
                     ${PROFILE_FIELDS}
                     ${!username ? USER_FIELDS : ""}
-                }
-            }`,
-            variables: { input }
-        });
-        return user;
-    }
-
-    /**
-     * Update a user profile.
-     * @param input Input arguments.
-     * @returns Updated user profile.
-     */
-    public async update (input: UpdateUserInput): Promise<User> {
-        const { data: { user } } = await this.client.query<{ user: User }>({
-            query: `mutation ($input: UpdateUserInput!) {
-                user: updateUser (input: $input) {
-                    ${PROFILE_FIELDS}
-                    ${USER_FIELDS}
                 }
             }`,
             variables: { input }